+++ src/main/java/ru/yandex/practicum/filmorate/storage/FeedStorage.java
@@ -7,7 +7,7 @@ import ru.yandex.practicum.filmorate.model.Feed;
 import java.util.List;
 
 public interface FeedStorage {
-    void addFeed(Integer entityId, Integer userId, long timeStamp, EventType eventType, OperationType operation);
+    void addFeed(Long entityId, Long userId, long timeStamp, EventType eventType, OperationType operation);
 
-    List<Feed> getFeed(int id);
+    List<Feed> getFeed(long id);
 }
