+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/FeedDbStorage.java
@@ -20,7 +20,7 @@ public class FeedDbStorage implements FeedStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void addFeed(Integer entityId, Integer userId, long timeStamp,
+    public void addFeed(Long entityId, Long userId, long timeStamp,
                         EventType eventType, OperationType operation) {
         String sqlQuery = "insert into events (entity_id, user_Id, time_stamp, event_type, event_operation) " +
                 "values (?, ?, ?, ?, ?)";
@@ -28,15 +28,15 @@ public class FeedDbStorage implements FeedStorage {
     }
 
     @Override
-    public List<Feed> getFeed(int id) {
+    public List<Feed> getFeed(long id) {
         String sqlGetFeed = "select * from events where user_id = ? order by EVENT_ID";
         return jdbcTemplate.query(sqlGetFeed, this::makeFeed, id);
     }
 
     Feed makeFeed(ResultSet rs, int rowNum) throws SQLException {
-        int eventId = rs.getInt("event_id");
-        int entityId = rs.getInt("entity_id");
-        int userId = rs.getInt("user_id");
+        long eventId = rs.getLong("event_id");
+        long entityId = rs.getLong("entity_id");
+        long userId = rs.getLong("user_id");
         long time = rs.getLong("time_stamp");
         EventType eventType = EventType.valueOf(rs.getString("event_type"));
         OperationType operation = OperationType.valueOf(rs.getString("event_operation"));
